refactor: HttpResponse를 추상클래스로 분리
